Move deferred tasks at the very end of the request and out of the request scope.
