+++ src/OrchardCore/OrchardCore.Modules/ModularTenantContainerMiddleware.cs
@@ -4,6 +4,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.DependencyInjection;
+using OrchardCore.DeferredTasks;
 using OrchardCore.Environment.Shell;
 
 namespace OrchardCore.Modules
@@ -33,21 +34,22 @@ namespace OrchardCore.Modules
             // Ensure all ShellContext are loaded and available.
             _orchardHost.Initialize();
 
-            var shellSetting = _runningShellTable.Match(httpContext);
+            var shellSettings = _runningShellTable.Match(httpContext);
 
             // Register the shell settings as a custom feature.
-            httpContext.Features.Set(shellSetting);
+            httpContext.Features.Set(shellSettings);
 
             // We only serve the next request if the tenant has been resolved.
-            if (shellSetting != null)
+            if (shellSettings != null)
             {
-                var shellContext = _orchardHost.GetOrCreateShellContext(shellSetting);
+                var shellContext = _orchardHost.GetOrCreateShellContext(shellSettings);
 
+                IDeferredTaskEngine deferredTaskEngine;
                 using (var scope = shellContext.EnterServiceScope())
                 {
                     if (!shellContext.IsActivated)
                     {
-                        var semaphore = _semaphores.GetOrAdd(shellSetting.Name, (name) => new SemaphoreSlim(1));
+                        var semaphore = _semaphores.GetOrAdd(shellSettings.Name, (name) => new SemaphoreSlim(1));
 
                         await semaphore.WaitAsync();
 
@@ -76,11 +78,27 @@ namespace OrchardCore.Modules
                         finally
                         {                            
                             semaphore.Release();
-                            _semaphores.TryRemove(shellSetting.Name, out semaphore);
+                            _semaphores.TryRemove(shellSettings.Name, out semaphore);
                         }
                     }
                     
                     await _next.Invoke(httpContext);
+                    deferredTaskEngine = scope.ServiceProvider.GetService<IDeferredTaskEngine>();
+                }
+
+                // Create a new scope only if there are pending tasks
+                if (deferredTaskEngine != null && deferredTaskEngine.HasPendingTasks)
+                {
+                    shellContext = _orchardHost.GetOrCreateShellContext(shellSettings);
+
+                    if (!shellContext.Released)
+                    {
+                        using (var scope = shellContext.EnterServiceScope())
+                        {
+                            var context = new DeferredTaskContext(scope.ServiceProvider);
+                            await deferredTaskEngine.ExecuteTasksAsync(context);
+                        }
+                    }
                 }
             }
         }
